# openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: API documentation for the Task Management web application, featuring task tracking, team collaboration, and real-time messaging.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base API path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
...
    TaskResponse:
      type: object
      properties:
        task_id:
          type: integer
        created_by:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        due_date:
          type: string
          format: date
        version:
          type: integer
    TaskStatus:
      type: string
      enum: [todo, in_progress, review, done, over_due]
    CreatedTaskRequest:
      type: object
      required:
        - title
        - due_date
      properties:
        team_id:
          type: integer
        title:
          type: string
        description:
          type: string
        due_date:
          type: string
          format: date
        owner:
          type: array
          items:
            type: string
            format: email
    UpdatedTaskRequest:
      type: object
      properties:
        new_due_date:
          type: string
          format: date
        new_status:
          $ref: '#/components/schemas/TaskStatus'
        version:
          type: integer
    TeamResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
    CreatedTeamRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
...
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        total_elements:
          type: integer
        page_number:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
paths:
...
  /api/v1/task:
    get:
      tags: [Tasks]
      summary: Filter tasks
      security:
        - BearerAuth: []
      parameters:
        - name: teamId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: first
          in: query
          schema:
            type: string
            format: date
        - name: second
          in: query
          schema:
            type: string
            format: date
        - name: sortBy
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - properties:
                      content:
                        type: array
                        items:
                          $ref: '#/components/schemas/TaskResponse'
    post:
      tags: [Tasks]
      summary: Create new task
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedTaskRequest'
      responses:
        '200':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /api/v1/task/unfinished:
    get:
      tags: [Tasks]
      summary: Get unfinished tasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of unfinished tasks
  /api/v1/task/{taskId}:
    patch:
      tags: [Tasks]
      summary: Update task status or due date
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatedTaskRequest'
      responses:
        '200':
          description: Task updated
  /api/v1/task/elastic-search:
    get:
      tags: [Tasks]
      summary: Advanced task search
      security:
        - BearerAuth: []
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Search results
  /api/v1/team:
    get:
      tags: [Teams]
      summary: Get all teams of user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - properties:
                      content:
                        type: array
                        items:
                          $ref: '#/components/schemas/TeamResponse'
    post:
      tags: [Teams]
      summary: Create team
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedTeamRequest'
      responses:
        '201':
          description: Team created
  /api/v1/team/{teamId}/members:
    get:
      tags: [Teams]
      summary: Get team members
      security:
        - BearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of members
  /api/v1/team/{teamId}/tasks:
    get:
      tags: [Teams]
      summary: Get tasks by team
      security:
        - BearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of tasks
  /api/v1/team-member:
    post:
      tags: [Teams]
      summary: Add member to team
      security:
        - BearerAuth: []
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: integer
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Member added
    patch:
      tags: [Teams]
      summary: Remove member from team
      security:
        - BearerAuth: []
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: integer
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member removed
...
  /api/v1/comment/task/{taskId}:
    get:
      tags: [Comments]
      summary: Get comments for a task
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of comments
...
  /api/v1/notification:
    get:
      tags: [Notifications]
      summary: Get all notifications
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of notifications
  /api/v1/notification/read-all:
    patch:
      tags: [Notifications]
      summary: Mark all as read
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
  /api/v1/notification/{notificationId}/read:
    patch:
      tags: [Notifications]
      summary: Mark specific notification as read
      security:
        - BearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
  /api/v1/notification/unread-count:
    get:
      tags: [Notifications]
      summary: Get unread count
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Count of unread notifications
...